pyproject.toml
[tool.poetry]
name = "tg-schedule-bot"
version = "0.1.0"
description = ""
authors = ["Vlatterran <soboleff@mail.ru>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
aiogram = "^3.13.0"
beautifulsoup4 = "^4.12.3"
typer = "^0.12.5"
lxml = "^5.3.0"
tortoise-orm = {extras = ["aiosqlite"], version = "^0.21.6"}

[tool.poetry.scripts]
schedule = 'tg_schedule_bot.cli:app'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
